+++ firmware/src/main.cpp
@@ -205,7 +205,7 @@ void handleSerial() {
       else if (serialBuf.startsWith("config ")) handleConfigCommand(serialBuf.substring(7));
       else if (serialBuf.startsWith("sens "))   handleSensCommand(serialBuf.substring(5));
       else if (serialBuf.startsWith("debug "))  handleDebugCommand(serialBuf.substring(6));
-      else if (serialBuf == "version")          { Serial.println("version=2.18.5"); }
+      else if (serialBuf == "version")          { Serial.println("version=2.19.0"); }
       else if (serialBuf.startsWith("volume ")) {
         int val = atoi(serialBuf.c_str() + 7);
         if (val >= 0 && val <= 100) {